# speed up `spec_is_possible()` helper